requirements.txt
# Core Dependencies (Lightweight)
groq==0.30.0
pypdf==5.9.0
python-dotenv==1.1.1

# Flask API Dependencies
flask==3.1.1
gunicorn==23.0.0
flask-cors==6.0.1

# Basic utilities
requests==2.32.4

# Lightweight text processing (no heavy ML models)
scikit-learn==1.3.0
